separate generators config and template to separate package
