+++ packages/graphql-codegen-compiler/src/clean-template.ts
