+++ packages/graphql-codegen-compiler/src/compile.ts
@@ -1,5 +1,6 @@
-import { GeneratorConfig, FileOutput, Settings, EInputType } from './types';
+import { FileOutput, Settings } from './types';
 import { debugLog, Document, Fragment, Operation, SchemaTemplateContext } from 'graphql-codegen-core';
+import { GeneratorConfig, EInputType } from 'graphql-codegen-generators';
 import { compile, registerPartial } from 'handlebars';
 import { initHelpers } from './handlebars-extensions';
 import { flattenTypes } from './flatten-types';
