+++ packages/graphql-codegen-compiler/src/flatten-types.ts
