+++ packages/graphql-codegen-compiler/src/generate-single-file.ts
@@ -1,6 +1,7 @@
 import { prepareSchemaForDocumentsOnly } from './prepare-documents-only';
-import { FileOutput, GeneratorConfig, Settings } from './types';
+import { FileOutput, Settings } from './types';
 import { SchemaTemplateContext, Document, debugLog } from 'graphql-codegen-core';
+import { GeneratorConfig } from 'graphql-codegen-generators';
 
 export function generateSingleFile(compiledIndexTemplate: HandlebarsTemplateDelegate, executionSettings: Settings, config: GeneratorConfig, templateContext: SchemaTemplateContext, documents: Document): FileOutput[] {
   debugLog(`[generateSingleFile] Compiling single file to: ${config.outFile}`);
