+++ packages/graphql-codegen-compiler/src/index.ts
@@ -1,0 +1,4 @@
+export { compileTemplate } from './compile';
+export { FileOutput } from './types';
+export { ALLOWED_CUSTOM_TEMPLATE_EXT } from './generate-multiple-files';
+export { getGeneratorConfig } from 'graphql-codegen-generators';
