+++ packages/graphql-codegen-compiler/src/prepare-documents-only.ts
