+++ packages/graphql-codegen-compiler/tslint.json
@@ -1,0 +1,61 @@
+{
+  "rules": {
+    "align": [true, "parameters"],
+    "class-name": true,
+    "curly": true,
+    "eofline": true,
+    "forin": true,
+    "indent": [true, "spaces"],
+    "label-position": true,
+    "max-line-length": [true, 180],
+    "member-access": false,
+    "member-ordering": [true,
+      "public-before-private",
+      "static-before-instance",
+      "variables-before-functions"
+    ],
+    "no-arg": true,
+    "no-bitwise": true,
+    "no-console": [true,
+      "debug",
+      "info",
+      "time",
+      "timeEnd",
+      "trace"
+    ],
+    "no-construct": true,
+    "no-debugger": true,
+    "no-duplicate-variable": true,
+    "no-empty": false,
+    "no-eval": true,
+    "no-inferrable-types": true,
+    "no-shadowed-variable": false,
+    "no-string-literal": false,
+    "no-switch-case-fall-through": true,
+    "no-trailing-whitespace": [true, "ignore-template-strings"],
+    "no-unused-expression": true,
+    "no-use-before-declare": false,
+    "no-var-keyword": false,
+    "object-literal-sort-keys": false,
+    "no-conditional-assignment": true,
+    "one-line": [true,
+      "check-open-brace",
+      "check-whitespace"
+    ],
+    "quotemark": [true, "single"],
+    "radix": false,
+    "semicolon": true,
+    "triple-equals": [true, "allow-null-check"],
+    "variable-name": [true,
+      "check-format",
+      "allow-leading-underscore"
+    ],
+    "whitespace": [true,
+      "check-branch",
+      "check-decl",
+      "check-operator",
+      "check-separator",
+      "check-type"
+    ]
+  }
+}
