+++ packages/graphql-codegen-generators/package.json
@@ -1,56 +1,22 @@
 {
   "name": "graphql-codegen-generators",
   "version": "0.7.0-alpha.383c3d35",
-  "description": "GraphQL code generator templates",
+  "description": "GraphQL code generator predefined templates",
   "license": "MIT",
   "scripts": {
     "test": "jest",
-    "pretest": "yarn build",
     "build": "webpack"
   },
-  "peerDependencies": {
-    "graphql-codegen-core": ">0.7.0-alpha.383c3d35"
-  },
   "devDependencies": {
-    "@types/change-case": "^2.3.1",
-    "@types/common-tags": "^1.2.5",
-    "@types/jest": "^20.0.2",
-    "@types/node": "7",
     "awesome-typescript-loader": "^3.2.1",
-    "dts-bundle": "^0.7.3",
-    "graphql": "^0.10.3",
-    "graphql-tag": "^2.4.2",
-    "graphql-tools": "^1.0.0",
-    "jest": "^20.0.4",
     "raw-loader": "^0.5.1",
-    "rimraf": "^2.6.1",
-    "tslint": "^5.4.3",
     "typescript": "^2.4.1",
-    "webpack": "^3.0.0",
-    "webpack-node-externals": "^1.6.0",
-    "graphql-codegen-core": ">0.7.0-alpha.0"
+    "webpack": "^3.2.0",
+    "webpack-node-externals": "^1.6.0"
   },
-  "dependencies": {
-    "@types/handlebars": "^4.0.33",
-    "change-case": "^3.0.1",
-    "common-tags": "^1.4.0",
-    "handlebars": "^4.0.10"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js"
-    ],
-    "transform": {
-      "^.+\\.(ts|tsx)$": "<rootDir>/tests/jest.typescript.js"
-    },
-    "testRegex": "tests/.*\\.ts$",
-    "testEnvironment": "node"
-  },
+  "main": "./dist/index.js",
   "typings": "dist/src/index.d.ts",
   "typescript": {
     "definition": "dist/src/index.d.ts"
-  },
-  "main": "./dist/index.js"
+  }
 }
