+++ packages/graphql-codegen-generators/src/index.ts
@@ -1,13 +1,10 @@
-export { compileTemplate } from './compile';
-export { GeneratorConfig, FileOutput, EInputType } from './types';
-export { ALLOWED_CUSTOM_TEMPLATE_EXT } from './generate-multiple-files';
-
 import TypescriptSingleFile from './typescript-single-file/config';
 import TypescriptMultiFile from './typescript-multi-file/config';
-
-export { getGeneratorConfig } from './get-generator';
 
 export {
   TypescriptSingleFile,
   TypescriptMultiFile,
 };
+
+export { GeneratorConfig, EInputType } from './types';
+export { getGeneratorConfig, definitions } from './get-generator';
