+++ packages/graphql-codegen-generators/src/types.ts
@@ -1,8 +1,3 @@
-import {
-  Operation, Document, Fragment, SelectionSetFragmentSpread,
-  SelectionSetFieldNode, SelectionSetInlineFragment
-} from 'graphql-codegen-core';
-
 export const EInputType = {
   SINGLE_FILE: 'SINGLE_FILE',
   MULTIPLE_FILES: 'MULTIPLE_FILES',
@@ -22,47 +17,4 @@ export interface GeneratorConfig {
   };
   outFile?: string;
   filesExtension?: string;
-}
-
-export interface FileOutput {
-  filename: string;
-  content: string;
-}
-
-export interface FlattenOperation extends Operation {
-  innerModels: FlattenModel[];
-  isFlatten: boolean;
-}
-
-export interface FlattenFragment extends Fragment {
-  innerModels: FlattenModel[];
-  isFlatten: boolean;
-}
-
-export interface FlattenDocument extends Document {
-  operations: FlattenOperation[];
-  fragments: FlattenFragment[];
-  hasFragments: boolean;
-  hasOperations: boolean;
-}
-
-export interface FlattenModel {
-  schemaBaseType: string;
-  modelType: string;
-  fields: SelectionSetFieldNode[];
-  fragmentsSpread: SelectionSetFragmentSpread[];
-  inlineFragments: SelectionSetInlineFragment[];
-  hasFragmentsSpread: boolean;
-  hasFields: boolean;
-  hasInlineFragments: boolean;
-}
-
-export interface Settings {
-  generateSchema?: boolean;
-  generateDocuments?: boolean;
-  verbose?: boolean;
-}
-
-export interface MultiFileTemplates {
-  [templateHandler: string]: Function[];
 }
